requirements.txt
# Core ML/DL Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.1.0
SimpleITK>=2.3.0
pydicom>=2.4.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Deep Learning Utilities
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# LLM and NLP
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
datasets>=2.12.0
tokenizers>=0.13.0

# Web Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
django-filter>=23.2

# Async Processing
celery>=5.3.0
redis>=4.6.0
kombu>=5.3.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# 3D Visualization
trimesh>=3.22.0
pygltflib>=1.15.0
open3d>=0.17.0

# Configuration and Utilities
PyYAML>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-django>=4.5.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Deployment
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0

# Monitoring and Logging
sentry-sdk>=1.30.0
structlog>=23.1.0

# ONNX Support
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0

# Additional Utilities
requests>=2.31.0
aiofiles>=23.2.0
python-json-logger>=2.0.0

